added contact me + pages refactor
added contact me component, navi, refactored some components into pages
