--- src/components/ContactMe.js
@@ -1,23 +1,0 @@
-
-
-const ContactMe = () => {
-
-  return (
-    <div className="App-header">
-      <h1>
-        About Me
-      </h1>
-      <p>
-        I am a student from The Bronx, New York attending Lehman College. I am a student of 
-        Prof. Flowers who teaches CMP 464. In this class we are learning react which is a
-        javascript library famously created and maintained by Facebook (now referred to as Meta).
-        I not only believe but understand that technology is the future and being apart of it makes
-        me happy.
-      </p>
-      <a href="https://github.com/iamblaw/"  target="_blank" > Click for Resume </a>
-    </div>
-
-  );
-}
- 
-export default ContactMe;
